backend/requirements.txt
django==4.2.17
djangorestframework==3.15.2
mongoengine==0.28.2
pymongo>=4.0,<4.7
dnspython==2.7.0
python-dotenv==1.0.1
django-cors-headers==4.6.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
